+++ langchain_hana/chains/graph_qa/__init__.py
@@ -1,3 +1,3 @@
-from .sparql_qa_chain import HanaSparqlQAChain
+from .hana_sparql_qa_chain import HanaSparqlQAChain
 
 __all__ = ["HanaSparqlQAChain"]
