+++ langchain_hana/chains/graph_qa/hana_sparql_qa_chain.py
